Auction Controller Bid Controller Updated Used Base Price

+++ server/src/main/java/com/project/dto/bids/BidReqDTO.java
@@ -14,5 +14,5 @@ public class BidReqDTO {
 
 	private Long auctionId;
     private Long userId;
-    private BigDecimal bidAmount;
+    private Double bidAmount;
 }
